// src/components/app_sidebar/types/app_sidebar.d.ts
interface SidebarItem {
  title: string;
  icon: React.ReactNode;
  onClick: () => void;
  color?:
    | "gray"
    | "gold"
    | "bronze"
    | "brown"
    | "yellow"
    | "amber"
    | "orange"
    | "tomato"
    | "red"
    | "ruby"
    | "crimson"
    | "pink"
    | "plum"
    | "purple"
    | "violet"
    | "iris"
    | "indigo"
    | "blue"
    | "cyan"
    | "teal"
    | "jade"
    | "green"
    | "grass"
    | "lime"
    | "mint"
    | "sky";
}

export interface AppSidebarProps {
  items: SidebarItem[];
}